+++ src/app/app.component.html
@@ -1,6 +1,1 @@
-<!--<app-todo-list></app-todo-list>-->
-
-<!--<router-outlet></router-outlet>-->
-<!--<app-toaster></app-toaster>-->
-
 <app-task-landing></app-task-landing>
