+++ src/app/components/task-landing/task-landing.component.html
@@ -1,16 +1,13 @@
 <div class="wrapper">
-  <header class="header">My header</header>
+  <h1 class="header">TaskBoard</h1>
 
   <aside class="sidebar">
-    <div>
-      <a [routerLink]="['/tasks']"> Task </a>
-    </div>
-    <div>
-      <a [routerLink]="['/board']"> board </a>
-    </div>
+    <mat-nav-list>
+      <a mat-list-item class="nav" [routerLink]="'/tasks'"> Task </a>
+      <a mat-list-item class="nav" [routerLink]="'/board'"> Board </a>
+    </mat-nav-list>
   </aside>
   <article class="content">
     <router-outlet></router-outlet>
   </article>
-  <!--  <app-todo-list class="content"></app-todo-list>-->
 </div>
