+++ src/app/components/task-landing/task-landing.component.scss
@@ -1,16 +1,4 @@
 HTML CSS Result Skip Results Iframe EDIT ON *,
-//*:before,
-//*:after {
-//  box-sizing: border-box;
-//}
-
-//body {
-//  margin: 40px;
-//  font-family: 'Open Sans', 'sans-serif';
-//  background-color: #fff;
-//  //color: #444;
-//}
-
 h1,
 p {
   margin: 0 0 1em 0;
@@ -46,17 +34,14 @@ p {
     grid-column: 1 / -1;
     /* needed for the floated layout */
     clear: both;
-  }
-  .header2 {
-    grid-column: 0 / 0;
-    /* needed for the floated layout */
-    clear: both;
+    font-weight: bold;
+    font-size: 3rem !important;
   }
 }
 
 .wrapper > * {
-  background-color: #eaac50;
-  color: #fff;
+  background-color: rgba(126, 232, 54, 0.37);
+  color: #223a41;
   border-radius: 5px;
   padding: 20px;
   font-size: 150%;
@@ -70,4 +55,11 @@ p {
     width: auto;
     margin: 0;
   }
+}
+
+.nav {
+  background-color: #9cc929;
+  margin: 10px;
+  border-radius: 10px;
+  width: 50%;
 }
