+++ src/app/features/task-board/components/task-board-column/task-board-column.component.scss
@@ -1,0 +1,66 @@
+.example-container {
+  width: 31%;
+  max-width: 100%;
+  margin: 0 25px 25px 0;
+  display: inline-block;
+  vertical-align: top;
+  height: 100%;
+}
+
+.example-list {
+  border: solid 1px #ccc;
+  background: white;
+  border-radius: 4px;
+  overflow: hidden;
+  display: block;
+  height: 100%;
+  min-height: 33vw;
+  margin: 5px;
+  padding: 5px 15px 5px 5px;
+}
+
+.example-box {
+  padding: 20px;
+  margin: 5px;
+  border-bottom: solid 1px #d54040;
+  color: rgba(0, 0, 0, 0.87);
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  //justify-content: space-between;
+  box-sizing: border-box;
+  cursor: move;
+  background: #73ecc6;
+  border-radius: 10px;
+  font-size: 14px;
+  align-content: center;
+  justify-items: center;
+  width: 100%;
+}
+
+.cdk-drag-preview {
+  box-sizing: border-box;
+  border-radius: 4px;
+  box-shadow: 0 5px 5px -3px rgba(0, 0, 0, 0.2), 0 8px 10px 1px rgba(0, 0, 0, 0.14), 0 3px 14px 2px rgba(0, 0, 0, 0.12);
+}
+
+.cdk-drag-placeholder {
+  opacity: 0;
+}
+
+.cdk-drag-animating {
+  transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
+}
+
+.example-box:last-child {
+  border: none;
+}
+
+.example-list.cdk-drop-list-dragging .example-box:not(.cdk-drag-placeholder) {
+  transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
+}
+
+.column-title {
+  display: flex;
+  justify-content: center;
+}
