+++ src/app/features/task-board/components/task-board/task-board.component.ts
@@ -1,7 +1,7 @@
 import { Component, inject, OnInit } from '@angular/core';
-import { CdkDragDrop } from '@angular/cdk/drag-drop';
-import { map, Observable } from 'rxjs';
-import { Todo, TodoStatus } from '@interfaces/Todo';
+
+import { map } from 'rxjs';
+import { TodoStatus } from '@interfaces/Todo';
 import { TodoStore } from '@services/todo-store.service';
 
 @Component({
@@ -23,24 +23,7 @@ export class TaskBoardComponent implements OnInit {
     map(todoList => todoList.filter(item => item.status === TodoStatus.Completed))
   );
 
-  drop(event: CdkDragDrop<Observable<Todo[]>>, newTaskStatus: TodoStatus) {
-    const id = event.item.element.nativeElement.getAttribute('taskid');
-    if (id) {
-      this.onItemStatusChanged([+id, newTaskStatus]);
-    }
-  }
-
   ngOnInit(): void {
     this.store.getTodos();
-  }
-
-  onItemStatusChanged = (changeParams: [number, TodoStatus]) => {
-    this.store.changeTodoStatus(changeParams[0], changeParams[1]).subscribe(_ => {
-      this.store.getTodos();
-    });
-  };
-
-  trackByItem(index: number, el: Todo): number {
-    return el.id;
   }
 }
