+++ src/app/features/task-board/task-board.module.ts
@@ -7,9 +7,10 @@ import { SharedModule } from '../@shared/shared.module';
 import { TaskBoardComponent } from './components/task-board/task-board.component';
 import { CdkDropList, DragDropModule } from '@angular/cdk/drag-drop';
 import { TaskBacklogModule } from '../task-backlog/task-backlog.module';
+import { TaskBoardColumnComponent } from './components/task-board-column/task-board-column.component';
 
 @NgModule({
-  declarations: [TaskBoardComponent],
+  declarations: [TaskBoardComponent, TaskBoardColumnComponent],
   imports: [
     CommonModule,
     TaskBacklogModule,
